moving aroudn mobile
