+++ themes/stanford_basic/algolia-search/src/search-box.tsx
@@ -1,11 +1,10 @@
 import {
   useClearRefinements,
-  useRefinementList,
   useSearchBox
 } from "react-instantsearch";
 import {useRef} from "preact/compat";
-import {CheckboxLabel, SearchForm, SearchInput} from "./styled-components";
-
+import {SearchForm, SearchInput} from "./styled-components";
+import RefinementSidebar from "./federated-search-facets";
 
 const SearchBox = ({federatedSearch}: { federatedSearch?: boolean }) => {
   const {query, refine} = useSearchBox();
@@ -73,51 +72,6 @@ const SearchBox = ({federatedSearch}: { federatedSearch?: boolean }) => {
   );
 }
 
-const ClearFilters = () => {
-  const { canRefine, refine } = useClearRefinements();
-
-  return (
-    <a
-      type="link"
-      onClick={refine}
-      className="clear-filters-link"
-    >
-      Reset filters
-    </a>
-  );
-};
-
-
-const RefinementSidebar = () => {
-  const {
-    items: sites,
-    refine: refineSites
-  } = useRefinementList({
-    attribute: "site_name",
-    limit: 100,
-    showMore: false,
-    sortBy: ["name"]
-  })
-
-  return (
-    <div className="federated-search-facets">
-      <h2>Filter by</h2>
-      <ClearFilters/>
-      <fieldset>
-        <legend>Sites</legend>
-        {sites.map(site =>
-          <CheckboxLabel key={site.label}>
-            <input type="checkbox" onChange={() => refineSites(site.value)}
-                   checked={site.isRefined}/>
-            <span className="checkbox">
-              <i class="fa-solid fa-check"></i>
-            </span>
-            <span className="label-display">{site.label} ({site.count})</span>
-          </CheckboxLabel>
-        )}
-      </fieldset>
-    </div>
-  )
-}
+<RefinementSidebar/>
 
 export default SearchBox;
